feat: add CMake build system and embed script
- CMakeLists.txt: CMake 3.20+, C11/C++17 standards
- CMakeLists.txt: auto-embed index.html via Python script
- scripts/embed_html.py: converts HTML to C string header
- transfer.c/.h: Phase 2 stub files for future file transfer
- .gitignore: ignore build/ directory and editor temp files
- Fix format warning in util.c
- Verified: clean build with zero warnings on macOS

+++ src/util.c
@@ -23,9 +23,9 @@ void util_log(LogLevel level, const char *fmt, ...)
     gettimeofday(&tv, NULL);
     struct tm *tm = localtime(&tv.tv_sec);
 
-    fprintf(stderr, "[%02d:%02d:%02d.%03ld] [%s] ",
+    fprintf(stderr, "[%02d:%02d:%02d.%03d] [%s] ",
             tm->tm_hour, tm->tm_min, tm->tm_sec,
-            tv.tv_usec / 1000, tag);
+            (int)(tv.tv_usec / 1000), tag);
 
     va_list ap;
     va_start(ap, fmt);
